Refactored bt_System class, created the Parallel_Threaded Node
Refactored the bt_System code so it now just looks after executing behavour trees.
Removed the bt_System from all of the nodes, this is was just a bad idea and not required. What I use to use it for it better to do as its own Node.
Used the bt-System class to manage all of the parallel execution engines. for the parallel_threaded class